cleaning and added tests
